Added weighted jobs to matrix
Now, the matrix can also show the number of emails sent per person of a certain job title. So we can now get the insight that employees send the most emails in total for example, but the presidents send more emails per person.
PS: The code is a huge mess still, I might clean it up later.